# Reduce allocations in cachecontrol parsing