```
benchmark                          old ns/op     new ns/op     delta
BenchmarkCacheControlParsing-4     3573          2119          -40.69%

benchmark                          old allocs     new allocs     delta
BenchmarkCacheControlParsing-4     61             4              -93.44%

benchmark                          old bytes     new bytes     delta
BenchmarkCacheControlParsing-4     976           432           -55.74%
```